permitir registro usuario

+++ src/Controller/EliminarTareaController.php
@@ -8,6 +8,7 @@ use Symfony\Component\Routing\Annotation\Route;
 class EliminarTareaController extends AbstractController{
     #[Route("/tareas/eliminar/{id}")]
     public function start(int $id, TareaRepository $tareaRepository){
+        $this->denyAccessUnlessGranted("IS_AUTHENTICATED_FULLY");
         $tarea = $tareaRepository->find($id);
         if($tarea){
             $tareaRepository->remove($tarea, true);
